Update data model
A couple of days ago, I received some feedback about the data model
from @aspacca.

After a brief discussion, we realized the current data model is only
adequate for some use cases.

Here's the current one:

{
    "function_id": "80036aa038",
    "function_version": "v0.0.0",
    "execution_id": "6b76b62589",
    "cloud_provider": "aws",
    "cloud_region": "us-east-1",
    "memory_limit_in_mb": "512",
    "input": {
        "type": "s3-sqs",
        "outputs": [
            "elasticsearch"
        ]
    },
}

As Andrea highlighted, there are circumstances (replay and continuing
queues) where a single lambda execution can involve multiple input
types. We need to have more than just a single input.

So I changed the data model to drop the `input` and `outputs` fields
and add a `inputs` list with all inputs involved in the execution.

And here's the updated data model:

{
    "function_id": "80036aa038",
    "function_version": "v0.0.0",
    "execution_id": "6b76b62589",  # optional?
    "cloud_provider": "aws",
    "cloud_region": "us-east-1",
    "memory_limit_in_mb": "512",
    "inputs": [
		{"type": "s3-sqs", "output": "elasticsearch"},
		{"type": "cloudwatch-logs", "output": "elasticsearch"}
	]
}

Besides the data model, we also noticed that the telemetry function
`input_selected_telemetry` can be called multiple times when replay
and continuing queues are involved.

To avoid multiple calls to the Telemetry API, we added a new telemetry
event called `event_processed_telemetry` triggered when the lambda
function has processed all events and is about to flush the shippers
to Elasticsearch.

The telemetry sends the data to the telemetry endpoint on the
`event_processed_telemetry` event.